requirement.txt
# Core dependencies for Week 1: Basic Dobot Arm Control
Jinja2==3.1.5
typeguard==4.4.2
pyserial==3.4
PyYAML==6.0.2
pydobotplus

# Note: Additional dependencies (numpy, opencv, pyrealsense2) will be needed 
# for advanced workshops involving computer vision and object detection